Rename storage structs more uniformly
