+++ src/backends.rs
@@ -1,6 +1,6 @@
+#[cfg(feature = "amazon-s3")]
+pub mod amazon_s3;
 #[cfg(feature = "google-drive")]
-pub mod googledrive;
+pub mod google_drive;
 #[cfg(feature = "localdisk")]
 pub mod localdisk;
-#[cfg(feature = "amazon-s3")]
-pub mod s3;
